refactor(audio): 音频内部数据统一 tick 域，audible_notes 内存降 1/3
- AudibleNote 存 start_tick/end_tick（u32）替代 start_sample/end_sample（u64）：1 亿音符 24→16 字节/条（-0.8GB），且模型桶天然有序，构建免 sort
- SortedCC 存 tick（u32）替代 sample（u64），flatten_automation_to_cc_events 不再做 tick→sample 转换
- dispatch/render/chase 全部 tick 域比较：dispatch 基准 = engine.current_tick（seek/块末同步），只有渲染段边界才转一次 sample（每块事件数量级）
- 新增 sample_to_tick（floor + 浮点误差向上校验），块边界反查不漏事件
- 引擎加 current_tick 字段，seek_to/load_model/apply_notes_only 同步维护；chase 命令 target 改 tick
- 预览：PreviewNoteParams 改 target_tick/duration_ticks，渲染线程转 sample 差喂预览引擎（内部不变）
- 测试换算到 tick 域（@44100：22050→480 tick 等）

+++ crates/yinhe-audio/src/export.rs
@@ -340,9 +340,8 @@ pub fn export_wav_gpu(
             }
 
             let start_sample =
-                crate::audio_model::tick_to_sample(note.start_tick as u64, segments, tpb, sr);
-            let end_sample =
-                crate::audio_model::tick_to_sample(note.end_tick as u64, segments, tpb, sr);
+                crate::audio_model::tick_to_sample(note.start_tick, segments, tpb, sr);
+            let end_sample = crate::audio_model::tick_to_sample(note.end_tick, segments, tpb, sr);
 
             events.push(yinhe_synth::SynthEvent {
                 sample: start_sample,
@@ -374,12 +373,7 @@ pub fn export_wav_gpu(
         let mut max_sample = 0u64;
         for key in 0..128usize {
             if let Some(last_note) = model.notes[key].last() {
-                let end = crate::audio_model::tick_to_sample(
-                    last_note.end_tick as u64,
-                    segments,
-                    tpb,
-                    sr,
-                );
+                let end = crate::audio_model::tick_to_sample(last_note.end_tick, segments, tpb, sr);
                 if end > max_sample {
                     max_sample = end;
                 }
